[#1975] Extract TxResubmissionAction's core into a shared TxResubmitter
TxResubmissionAction's prune/find/throttle/resubmit logic moves verbatim
into a new engine-independent TxResubmitter (Transaction/TxResubmitter.swift)
so a later task can drive the same resubmission core from
SlipstreamSynchronizer. TxResubmissionAction becomes a thin Action adapter:
it owns a TxResubmitter, extracts latestBlockHeight from the ActionContext,
and delegates to checkAndResubmit(latestBlockHeight:), keeping only the
state-machine plumbing (removeBlocksCacheWhenFailed, the prevState-driven
transition, stop()).

Two forwarding computed properties keep the action's existing external
surface working unchanged: latestResolvedTime (poked directly by
TxResubmissionActionTests) and transactionEncoder (re-wired by
CompactBlockProcessor.updateService(_:) after a server switch).

No behavior change; TxResubmissionActionTests is unmodified and green.
